Quick code cleanup and updates
No structural changes yet. Just trying to get it to run solidly and then port to SDK4. Currently looks fine, but the buttons don't work and trying to access the menu, health, or timeline causes the app to crash hard. Probably more functions going on with regard to the "gameplay" aspect of the watch face than I'll end up using, and simplifying may help weed out the bugs.